requirements.txt
click==7.1.2
Flask==1.1.2
flask-mongoengine==0.9.5
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.20.0
pymongo==3.10.1
pyspellchecker==0.5.4
six==1.14.0
Unidecode==1.1.1
Werkzeug==1.0.1
WTForms==2.3.1
#production_staging environment
gunicorn